Fix bug with incorrect timestamps
When a whole new block of recent readings are read, they got spaced
with 15 minutes (like the history readings), instead of with 1 minute.

Fixed in DataContainer.

Modified ManualGlucoseReading to not autogenerate.
Did not bother to add a migration, so having an old database will
crash the app. Share it with yourself, and wipe app data.